@@ -18,13 +18,13 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_markdown: ^0.6.1
-  flutter_svg: ^0.19.3
+  flutter_svg: ^0.21.0
 
 
   # The following adds the Cupertino Icons font to your application.
