@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 
 class Button extends StatelessWidget {
   const Button({
-    @required this.onPress,
-    @required this.title,
-    Key key,
+    required this.onPress,
+    required this.title,
+    Key? key,
   }) : super(key: key);
   
   final Function onPress;
@@ -13,7 +13,7 @@ class Button extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return ElevatedButton(
-      onPressed: onPress,
+      onPressed: onPress as void Function()?,
       child: Text(title),
     );
   }
