@@ -2,11 +2,11 @@ import 'package:flutter/material.dart';
 
 class PodcastImage extends StatelessWidget {
   const PodcastImage({
-    @required this.url,
-    Key key,
+    required this.url,
+    Key? key,
   }) : super(key: key);
 
-  final String url;
+  final String? url;
 
   @override
   Widget build(BuildContext context) {
@@ -14,7 +14,7 @@ class PodcastImage extends StatelessWidget {
       margin: const EdgeInsets.only(left: 16, top: 16),
       child: CircleAvatar(
         radius: 22,
-        backgroundImage: NetworkImage(url),
+        backgroundImage: NetworkImage(url!),
       ),
     );
   }
