@@ -2,16 +2,16 @@ import 'package:flutter/material.dart';
 
 class PostTitle extends StatelessWidget {
   const PostTitle({
-    @required this.title,
-    Key key,
+    required this.title,
+    Key? key,
   }) : super(key: key);
 
-  final String title;
+  final String? title;
 
   @override
   Widget build(BuildContext context) {
     return Text(
-      title,
+      title!,
       style: Theme.of(context).textTheme.headline3,
     );
   }
