@@ -3,24 +3,24 @@ import 'package:flutter/material.dart';
 
 class TagTitle extends StatelessWidget {
   const TagTitle({
-    @required this.tagName,
-    Key key,
+    required this.tagName,
+    Key? key,
   }) : super(key: key);
 
-  final String tagName;
+  final String? tagName;
 
   @override
   Widget build(BuildContext context) {
     return RichText(
       text: TextSpan(
         text: '#',
-        style: Theme.of(context).textTheme.headline2.copyWith(
+        style: Theme.of(context).textTheme.headline2!.copyWith(
           color: CustomColors.tagLabel,
         ),
         children: [
           TextSpan(
             text: tagName,
-            style: Theme.of(context).textTheme.headline2.copyWith(
+            style: Theme.of(context).textTheme.headline2!.copyWith(
               color: CustomColors.primaryText,
             ),
           )
