@@ -2,13 +2,13 @@ import 'package:flutter/material.dart';
 
 class VideoPreview extends StatelessWidget {
   const VideoPreview({
-    @required this.imageUrl,
-    @required this.duration,
-    Key key,
+    required this.imageUrl,
+    required this.duration,
+    Key? key,
   }) : super(key: key);
 
-  final String imageUrl;
-  final String duration;
+  final String? imageUrl;
+  final String? duration;
 
   @override
   Widget build(BuildContext context) {
@@ -16,7 +16,7 @@ class VideoPreview extends StatelessWidget {
       alignment: AlignmentDirectional.bottomEnd,
       children: [
         Image(
-          image: NetworkImage(imageUrl)
+          image: NetworkImage(imageUrl!)
         ),
         Container(
           padding: const EdgeInsets.fromLTRB(5, 2, 5, 3),
@@ -26,8 +26,8 @@ class VideoPreview extends StatelessWidget {
             color: const Color.fromRGBO(0, 0, 0, 0.8),
           ),
           child: Text(
-            duration,
-            style: Theme.of(context).textTheme.bodyText1.copyWith(
+            duration!,
+            style: Theme.of(context).textTheme.bodyText1!.copyWith(
               color: Colors.white,
             ),
           ),
