@@ -13,7 +13,7 @@ void main() {
 
 class MyApp extends StatelessWidget {
   const MyApp (
-    {Key key,}
+    {Key? key,}
   ) : super(key: key);
 
   @override
